Struct dioxus_core::prelude::VNode

pub struct VNode { /* private fields */ }

A reference to a template along with any context needed to hydrate it

The dynamic parts of the template are stored separately from the static parts. This allows faster diffing by skipping static parts of the template.

Implementations

impl VNode

pub fn empty() -> Element

Create a template with no nodes that will be skipped over during diffing

pub fn placeholder() -> Self

Create a template with a single placeholder node

pub fn new( key: Option<String>, template: Template, dynamic_nodes: Box<[DynamicNode]>, dynamic_attrs: Box<[Box<[Attribute]>]> ) -> Self

Create a new VNode

pub fn dynamic_root(&self, idx: usize) -> Option<&DynamicNode>

Load a dynamic root at the given index

Returns None if the root is actually a static node (Element/Text)

pub fn mounted_dynamic_node( &self, dynamic_node_idx: usize, dom: &VirtualDom ) -> Option<ElementId>

Get the mounted id for a dynamic node index

pub fn mounted_root( &self, root_idx: usize, dom: &VirtualDom ) -> Option<ElementId>

Get the mounted id for a root node index

pub fn mounted_dynamic_attribute( &self, dynamic_attribute_idx: usize, dom: &VirtualDom ) -> Option<ElementId>

Get the mounted id for a dynamic attribute index