dioxus_core::prelude

Struct VirtualDom

pub struct VirtualDom { /* private fields */ }

A virtual node system that progresses user events and diffs UI trees.

Guide

Components are defined as simple functions that take crate::properties::Properties and return an Element.

#[derive(Props, PartialEq, Clone)]
struct AppProps {
    title: String
}

fn app(cx: AppProps) -> Element {
    rsx!(
        div {"hello, {cx.title}"}
    )
}

Components may be composed to make complex apps.

static ROUTES: &str = "";

#[component]
fn app(cx: AppProps) -> Element {
    rsx!(
        NavBar { routes: ROUTES }
        Title { "{cx.title}" }
        Footer {}
    )
}

#[component]
fn NavBar( routes: &'static str) -> Element {
    rsx! {
        div { "Routes: {routes}" }
    }
}

#[component]
fn Footer() -> Element {
    rsx! { div { "Footer" } }
}

#[component]
fn Title( children: Element) -> Element {
    rsx! {
        div { id: "title", {children} }
    }
}

To start an app, create a VirtualDom and call VirtualDom::rebuild to get the list of edits required to draw the UI.

let mut vdom = VirtualDom::new(app);
let edits = vdom.rebuild_to_vec();

To call listeners inside the VirtualDom, call Runtime::handle_event with the appropriate event data.

let event = Event::new(std::rc::Rc::new(0) as std::rc::Rc<dyn std::any::Any>, true);
runtime.handle_event("onclick", event, ElementId(0));

While no events are ready, call VirtualDom::wait_for_work to poll any futures inside the VirtualDom.

tokio::runtime::Runtime::new().unwrap().block_on(async {
    vdom.wait_for_work().await;
});

Once work is ready, call VirtualDom::render_immediate to compute the differences between the previous and current UI trees. This will write edits to a WriteMutations object you pass in that contains with edits that need to be handled by the renderer.

let mut mutations = Mutations::default();

vdom.render_immediate(&mut mutations);

To not wait for suspense while diffing the VirtualDom, call VirtualDom::render_immediate.

Building an event loop around Dioxus:

Putting everything together, you can build an event loop around Dioxus by using the methods outlined above.

let mut real_dom = RealDom::new();

#[component]
fn app() -> Element {
    rsx! {
        div { "Hello World" }
    }
}

let mut dom = VirtualDom::new(app);

dom.rebuild(&mut real_dom.apply());

loop {
    tokio::select! {
        _ = dom.wait_for_work() => {}
        evt = real_dom.wait_for_event() => {
            let evt = dioxus_core::Event::new(evt, true);
            dom.runtime().handle_event("onclick", evt, ElementId(0))
        },
    }

    dom.render_immediate(&mut real_dom.apply());
}

Waiting for suspense

Because Dioxus supports suspense, you can use it for server-side rendering, static site generation, and other use cases where waiting on portions of the UI to finish rendering is important. To wait for suspense, use the VirtualDom::wait_for_suspense method:

tokio::runtime::Runtime::new().unwrap().block_on(async {
    let mut dom = VirtualDom::new(app);

    dom.rebuild_in_place();
    dom.wait_for_suspense().await;
});

// Render the virtual dom

Implementations

impl VirtualDom

pub fn new(app: fn() -> Element) -> Self

Create a new VirtualDom with a component that does not have special props.

Description

Later, the props can be updated by calling “update” with a new set of props, causing a set of re-renders.

This is useful when a component tree can be driven by external state (IE SSR) but it would be too expensive to toss out the entire tree.

Example

fn Example() -> Element  {
    rsx!( div { "hello world" } )
}

let dom = VirtualDom::new(Example);

Note: the VirtualDom is not progressed, you must either “run_with_deadline” or use “rebuild” to progress it.

pub fn new_with_props<P: Clone + 'static, M: 'static>( root: impl ComponentFunction<P, M>, root_props: P, ) -> Self

Create a new VirtualDom with the given properties for the root component.

Description

Later, the props can be updated by calling “update” with a new set of props, causing a set of re-renders.

This is useful when a component tree can be driven by external state (IE SSR) but it would be too expensive to toss out the entire tree.

Example

#[derive(PartialEq, Props, Clone)]
struct SomeProps {
    name: &'static str
}

fn Example(cx: SomeProps) -> Element  {
    rsx! { div { "hello {cx.name}" } }
}

let dom = VirtualDom::new_with_props(Example, SomeProps { name: "world" });

Note: the VirtualDom is not progressed on creation. You must either “run_with_deadline” or use “rebuild” to progress it.

let mut dom = VirtualDom::new_with_props(Example, SomeProps { name: "jane" });
dom.rebuild_in_place();

pub fn prebuilt(app: fn() -> Element) -> Self

Create a new virtualdom and build it immediately

pub fn get_scope(&self, id: ScopeId) -> Option<&ScopeState>

Get the state for any scope given its ID

This is useful for inserting or removing contexts from a scope, or rendering out its root node

pub fn base_scope(&self) -> &ScopeState

Get the single scope at the top of the VirtualDom tree that will always be around

This scope has a ScopeId of 0 and is the root of the tree

pub fn in_runtime<O>(&self, f: impl FnOnce() -> O) -> O

Run a closure inside the dioxus runtime

pub fn with_root_context<T: Clone + 'static>(self, context: T) -> Self

Build the virtualdom with a global context inserted into the base scope

This is useful for what is essentially dependency injection when building the app

pub fn provide_root_context<T: Clone + 'static>(&self, context: T)

Provide a context to the root scope

pub fn insert_any_root_context(&mut self, context: Box<dyn Any>)

Build the virtualdom with a global context inserted into the base scope

This method is useful for when you want to provide a context in your app without knowing its type

pub fn mark_dirty(&mut self, id: ScopeId)

Manually mark a scope as requiring a re-render

Whenever the Runtime “works”, it will re-render this scope

pub async fn wait_for_work(&mut self)

Wait for the scheduler to have any work.

This method polls the internal future queue, waiting for suspense nodes, tasks, or other work. This completes when any work is ready. If multiple scopes are marked dirty from a task or a suspense tree is finished, this method will exit.

This method is cancel-safe, so you’re fine to discard the future in a select block.

This lets us poll async tasks and suspended trees during idle periods without blocking the main thread.

Example

let dom = VirtualDom::new(app);

pub fn process_events(&mut self)

Process all events in the queue until there are no more left

pub fn rebuild_in_place(&mut self)

Rebuild the virtualdom without handling any of the mutations

This is useful for testing purposes and in cases where you render the output of the virtualdom without handling any of its mutations.

pub fn rebuild_to_vec(&mut self) -> Mutations

VirtualDom::rebuild to a vector of mutations for testing purposes

pub fn rebuild(&mut self, to: &mut impl WriteMutations)

Performs a full rebuild of the virtual dom, returning every edit required to generate the actual dom from scratch.

The mutations item expects the RealDom’s stack to be the root of the application.

Tasks will not be polled with this method, nor will any events be processed from the event queue. Instead, the root component will be run once and then diffed. All updates will flow out as mutations.

All state stored in components will be completely wiped away.

Any templates previously registered will remain.

Example

fn app() -> Element {
    rsx! { "hello world" }
}

let mut dom = VirtualDom::new(app);
let mut mutations = Mutations::default();
dom.rebuild(&mut mutations);

pub fn render_immediate(&mut self, to: &mut impl WriteMutations)

Render whatever the VirtualDom has ready as fast as possible without requiring an executor to progress suspended subtrees.

pub fn render_immediate_to_vec(&mut self) -> Mutations

Self::render_immediate to a vector of mutations for testing purposes

pub async fn wait_for_suspense(&mut self)

Render the virtual dom, waiting for all suspense to be finished

The mutations will be thrown out, so it’s best to use this method for things like SSR that have async content

We don’t call “flush_sync” here since there’s no sync work to be done. Futures will be progressed like usual, however any futures waiting on flush_sync will remain pending

pub fn suspended_tasks_remaining(&self) -> bool

Check if there are any suspended tasks remaining

pub async fn wait_for_suspense_work(&mut self)

Wait for the scheduler to have any work that should be run during suspense.

pub async fn render_suspense_immediate(&mut self) -> Vec<ScopeId>

Render any dirty scopes immediately, but don’t poll any futures that are client only on that scope Returns a list of suspense boundaries that were resolved

pub fn runtime(&self) -> Rc<Runtime>

Get the current runtime

pub fn handle_event( &self, name: &str, event: Rc<dyn Any>, element: ElementId, bubbling: bool, )

👎Deprecated: Use [VirtualDom::runtime().handle_event] instead

Handle an event with the Virtual Dom. This method is deprecated in favor of [VirtualDom::runtime().handle_event] and will be removed in a future release.