Struct dioxus_core::Scoped

pub struct Scoped<'a, T = ()> {
    pub scope: &'a ScopeState,
    pub props: &'a T,
}

A wrapper around a component’s ScopeState and properties. The ScopeState provides the majority of methods for the VirtualDom and component state.

Fields
scope: &'a ScopeState

The component’s state and handle to the scheduler.

Stores things like the custom bump arena, spawn functions, hooks, and the scheduler.

props: &'a T

The component’s properties.