Struct dioxus_core::VirtualDom

pub struct VirtualDom { /* private fields */ }

A virtual node system that progresses user events and diffs UI trees.

Guide

Components are defined as simple functions that take Scope and return an Element.

#[derive(Props, PartialEq)]
struct AppProps {
    title: String
}

fn App(cx: Scope<AppProps>) -> Element {
    cx.render(rsx!(
        div {"hello, {cx.props.title}"}
    ))
}

Components may be composed to make complex apps.

static ROUTES: &str = "";

#[component]
fn App(cx: Scope<AppProps>) -> Element {
    cx.render(rsx!(
        NavBar { routes: ROUTES }
        Title { "{cx.props.title}" }
        Footer {}
    ))
}

#[component]
fn NavBar(cx: Scope, routes: &'static str) -> Element {
    cx.render(rsx! {
        div { "Routes: {routes}" }
    })
}

#[component]
fn Footer(cx: Scope) -> Element {
    cx.render(rsx! { div { "Footer" } })
}

#[component]
fn Title<'a>(cx: Scope<'a>, children: Element<'a>) -> Element {
    cx.render(rsx! {
        div { id: "title", children }
    })
}

To start an app, create a VirtualDom and call VirtualDom::rebuild to get the list of edits required to draw the UI.

let mut vdom = VirtualDom::new(App);
let edits = vdom.rebuild();

To call listeners inside the VirtualDom, call VirtualDom::handle_event with the appropriate event data.

vdom.handle_event(event);

While no events are ready, call VirtualDom::wait_for_work to poll any futures inside the VirtualDom.

vdom.wait_for_work().await;

Once work is ready, call VirtualDom::render_with_deadline to compute the differences between the previous and current UI trees. This will return a Mutations object that contains Edits, Effects, and NodeRefs that need to be handled by the renderer.

let mutations = vdom.work_with_deadline(tokio::time::sleep(Duration::from_millis(100)));

for edit in mutations.edits {
    real_dom.apply(edit);
}

To not wait for suspense while diffing the VirtualDom, call VirtualDom::render_immediate or pass an immediately ready future to VirtualDom::render_with_deadline.

Building an event loop around Dioxus:

Putting everything together, you can build an event loop around Dioxus by using the methods outlined above.

#[component]
fn App(cx: Scope) -> Element {
    cx.render(rsx! {
        div { "Hello World" }
    })
}

let dom = VirtualDom::new(App);

real_dom.apply(dom.rebuild());

loop {
    select! {
        _ = dom.wait_for_work() => {}
        evt = real_dom.wait_for_event() => dom.handle_event(evt),
    }

    real_dom.apply(dom.render_immediate());
}

Waiting for suspense

Because Dioxus supports suspense, you can use it for server-side rendering, static site generation, and other usecases where waiting on portions of the UI to finish rendering is important. To wait for suspense, use the VirtualDom::render_with_deadline method:

let dom = VirtualDom::new(app);

let deadline = tokio::time::sleep(Duration::from_millis(100));
let edits = dom.render_with_deadline(deadline).await;

Use with streaming

If not all rendering is done by the deadline, it might be worthwhile to stream the rest later. To do this, we suggest rendering with a deadline, and then looping between VirtualDom::wait_for_work and render_immediate until no suspended work is left.

let dom = VirtualDom::new(app);

let deadline = tokio::time::sleep(Duration::from_millis(20));
let edits = dom.render_with_deadline(deadline).await;

real_dom.apply(edits);

while dom.has_suspended_work() {
   dom.wait_for_work().await;
   real_dom.apply(dom.render_immediate());
}

Implementations

impl VirtualDom

pub fn new(app: fn(_: Scope<'_>) -> Element<'_>) -> Self

Create a new VirtualDom with a component that does not have special props.

Description

Later, the props can be updated by calling “update” with a new set of props, causing a set of re-renders.

This is useful when a component tree can be driven by external state (IE SSR) but it would be too expensive to toss out the entire tree.

Example

fn Example(cx: Scope) -> Element  {
    cx.render(rsx!( div { "hello world" } ))
}

let dom = VirtualDom::new(Example);

Note: the VirtualDom is not progressed, you must either “run_with_deadline” or use “rebuild” to progress it.

pub fn new_with_props<P: 'static>( root: fn(_: Scope<'_, P>) -> Element<'_>, root_props: P ) -> Self

Create a new VirtualDom with the given properties for the root component.

Description

Later, the props can be updated by calling “update” with a new set of props, causing a set of re-renders.

This is useful when a component tree can be driven by external state (IE SSR) but it would be too expensive to toss out the entire tree.

Example

#[derive(PartialEq, Props)]
struct SomeProps {
    name: &'static str
}

fn Example(cx: Scope<SomeProps>) -> Element  {
    cx.render(rsx!{ div{ "hello {cx.props.name}" } })
}

let dom = VirtualDom::new(Example);

Note: the VirtualDom is not progressed on creation. You must either “run_with_deadline” or use “rebuild” to progress it.

let mut dom = VirtualDom::new_with_props(Example, SomeProps { name: "jane" });
let mutations = dom.rebuild();

pub fn get_scope(&self, id: ScopeId) -> Option<&ScopeState>

Get the state for any scope given its ID

This is useful for inserting or removing contexts from a scope, or rendering out its root node

pub fn base_scope(&self) -> &ScopeState

Get the single scope at the top of the VirtualDom tree that will always be around

This scope has a ScopeId of 0 and is the root of the tree

pub fn with_root_context<T: Clone + 'static>(self, context: T) -> Self

Build the virtualdom with a global context inserted into the base scope

This is useful for what is essentially dependency injection when building the app

pub fn mark_dirty(&mut self, id: ScopeId)

Manually mark a scope as requiring a re-render

Whenever the Runtime “works”, it will re-render this scope

pub fn handle_event( &mut self, name: &str, data: Rc<dyn Any>, element: ElementId, bubbles: bool )

Call a listener inside the VirtualDom with data from outside the VirtualDom.

This method will identify the appropriate element. The data must match up with the listener delcared. Note that this method does not give any indication as to the success of the listener call. If the listener is not found, nothing will happen.

It is up to the listeners themselves to mark nodes as dirty.

If you have multiple events, you can call this method multiple times before calling “render_with_deadline”

pub async fn wait_for_work(&mut self)

Wait for the scheduler to have any work.

This method polls the internal future queue, waiting for suspense nodes, tasks, or other work. This completes when any work is ready. If multiple scopes are marked dirty from a task or a suspense tree is finished, this method will exit.

This method is cancel-safe, so you’re fine to discard the future in a select block.

This lets us poll async tasks and suspended trees during idle periods without blocking the main thread.

Example

let dom = VirtualDom::new(App);
let sender = dom.get_scheduler_channel();

pub fn process_events(&mut self)

Process all events in the queue until there are no more left

pub fn replace_template(&mut self, template: Template<'static>)

Replace a template at runtime. This will re-render all components that use this template. This is the primitive that enables hot-reloading.

The caller must ensure that the template refrences the same dynamic attributes and nodes as the original template.

This will only replace the the parent template, not any nested templates.

pub fn rebuild(&mut self) -> Mutations<'_>

Performs a full rebuild of the virtual dom, returning every edit required to generate the actual dom from scratch.

The mutations item expects the RealDom’s stack to be the root of the application.

Tasks will not be polled with this method, nor will any events be processed from the event queue. Instead, the root component will be ran once and then diffed. All updates will flow out as mutations.

All state stored in components will be completely wiped away.

Any templates previously registered will remain.

Example

static App: Component = |cx|  cx.render(rsx!{ "hello world" });

let mut dom = VirtualDom::new();
let edits = dom.rebuild();

apply_edits(edits);

pub fn render_immediate(&mut self) -> Mutations<'_>

Render whatever the VirtualDom has ready as fast as possible without requiring an executor to progress suspended subtrees.

pub async fn wait_for_suspense(&mut self)

Render the virtual dom, waiting for all suspense to be finished

The mutations will be thrown out, so it’s best to use this method for things like SSR that have async content

pub async fn render_with_deadline( &mut self, deadline: impl Future<Output = ()> ) -> Mutations<'_>

Render what you can given the timeline and then move on

It’s generally a good idea to put some sort of limit on the suspense process in case a future is having issues.

If no suspense trees are present

pub fn runtime(&self) -> Rc<Runtime>

Get the current runtime