dioxus_core

Trait IntoDynNode

pub trait IntoDynNode<A = ()> {
    // Required method
    fn into_dyn_node(self) -> DynamicNode;
}

A trait that allows various items to be converted into a dynamic node for the rsx macro

Required Methods

fn into_dyn_node(self) -> DynamicNode

Consume this item and produce a DynamicNode