Type Alias dioxus_core::Element

pub type Element<'a> = Option<VNode<'a>>;

An Element is a possibly-none VNode created by calling render on Scope or ScopeState.

An Errored Element will propagate the error to the nearest error boundary.

Aliased Type

enum Element<'a> {
    None,
    Some(VNode<'a>),
}

Variants

None

No value.

Some(VNode<'a>)

Some value of type T.