dioxus_core

Type Alias Result

pub type Result<T = ()> = Result<T, CapturedError>;

A type alias for a result that can be either a boxed error or a value This is useful to avoid having to use Result<T, CapturedError> everywhere

Aliased Type

enum Result<T = ()> {
    Ok(T),
    Err(CapturedError),
}

Variants

Ok(T)

Contains the success value

Err(CapturedError)

Contains the error value