dioxus_core

Module prelude

The purpose of this module is to alleviate imports of many common types

This includes types like Element, and Component.

Re-exports

pub use crate::innerlude::AnyValue;

pub use crate::innerlude::OptionStringFromMarker;

pub use crate::innerlude::SuspenseBoundary;

pub use crate::innerlude::SuspenseBoundary;

Structs

Attribute

An attribute on a DOM node, such as id="my-thing" or href="https://example.com"

Callback

The callback type generated by the rsx! macro when an on field is specified for components.

ErrorContext

A context with information about suspended components

Event

A wrapper around some generic data that handles the event’s state

ReactiveContext

Reactivity

Runtime

A global runtime that is shared across all scopes that provides the async runtime and context API

RuntimeGuard

A guard for a new runtime. This must be used to override the current runtime when importing components from a dynamic library that has it’s own runtime.

ScopeId

A component’s unique identifier.

ScopeState

A component’s rendered state.

SuspendedFuture

A task that has been suspended which may have an optional loading placeholder

SuspenseBoundaryProps

Properties for the SuspenseBoundary() component.

SuspenseContext

A context with information about suspended components

Task

A task’s unique identifier.

Template

A static layout of a UI tree that describes a set of dynamic and static nodes.

VNode

A reference to a template along with any context needed to hydrate it

VNodeInner

A reference to a template along with any context needed to hydrate it

VirtualDom

A virtual node system that progresses user events and diffs UI trees.

Enums

RenderError

An error that can occur while rendering a component

TemplateAttribute

An attribute of the TemplateNode, created at compile time

TemplateNode

A statically known node in a layout.

Traits

ComponentFunction

Any component that implements the ComponentFn trait can be used as a component.

Context

Provides context methods to Result and Option types that are compatible with CapturedError

HasAttributes

A trait for anything that has a dynamic list of attributes

IntoAttributeValue

A value that can be converted into an attribute value

IntoDynNode

A trait that allows various items to be converted into a dynamic node for the rsx macro

Properties

Every “Props” used for a component must implement the Properties trait. This trait gives some hints to Dioxus on how to memoize the props and some additional optimizations that can be made. We strongly encourage using the derive macro to implement the Properties trait automatically.

SuperFrom

A enhanced version of the From trait that allows with more flexibility.

SuperInto

A enhanced version of the Into trait that allows with more flexibility.

SuspenseExtension

Provides context methods to Result<T, RenderError> to show loading indicators for suspended results

Functions

ErrorBoundary

Create a new error boundary component that catches any errors thrown from child components

Fragment

Create inline fragments using Component syntax.

SuspenseBoundary

Suspense Boundaries let you render a fallback UI while a child component is suspended.

consume_context

Consume context from the current scope

consume_context_from_scope

Consume context from the current scope

current_owner

Returns the current owner. This owner will be used to drop any Copy state that is created by the generational-box crate.

current_scope_id

Get the current scope id

fc_to_builder

This utility function launches the builder method so that the rsx! macro can use the typed-builder pattern to initialize a component’s props.

generation

Get the current render since the inception of this component

has_context

Check if the current scope has a context

needs_update

Mark the current scope as dirty, causing it to re-render

needs_update_any

Mark the current scope as dirty, causing it to re-render

parent_scope

Get the parent of the current scope if it exists

provide_context

Provide context to the current scope

provide_error_boundary

Provide an error boundary to catch errors from child components

provide_root_context

Provide a context to the root scope

queue_effect

Queue an effect to run after the next render. You generally shouldn’t need to interact with this function directly. use_effect will call this function for you.

remove_future

Informs the scheduler that this task is no longer needed and should be removed.

schedule_update

Schedule an update for the current component

schedule_update_any

Schedule an update for any component given its ScopeId.

spawn

Spawns the future but does not return the Task. This task will automatically be canceled when the component is dropped.

spawn_forever

Spawn a future that Dioxus won’t clean up when this component is unmounted

spawn_isomorphic

Start a new future on the same thread as the rest of the VirtualDom.

suspend

Suspended the current component on a specific task and then return None

throw_error

Throw a CapturedError into the current scope. The error will bubble up to the nearest crate::prelude::ErrorBoundary() or the root of the app.

try_consume_context

Consume context from the current scope

use_after_render

Push this function to be run after the next render

use_before_render

A hook that allows you to insert a “before render” function.

use_drop

Creates a callback that will be run before the component is removed. This can be used to clean up side effects from the component (created with use_effect).

use_hook

Store a value between renders. The foundational hook for all other hooks.

use_hook_with_cleanup

Use a hook with a cleanup function

with_owner

Run a closure with the given owner.

Type Aliases

Component

A Component is a function that takes Properties and returns an Element.

Element

An Element is a possibly-none VNode created by calling render on ScopeId or ScopeState.

EventHandler

The callback type generated by the rsx! macro when an on field is specified for components.