Trait dioxus_core::prelude::Throw

pub trait Throw<S = ()>: Sized {
    type Out;

    // Required methods
    fn throw(self, cx: &ScopeState) -> Option<Self::Out>;
    fn throw_with<D: Debug + 'static>(
        self,
        cx: &ScopeState,
        e: impl FnOnce() -> D
    ) -> Option<Self::Out>;
}

A trait to allow results to be thrown upwards to the nearest Error Boundary

The canonical way of using this trait is to throw results from hooks, aborting rendering through question mark syntax. The throw method returns an option that evaluates to None if there is an error, injecting the error to the nearest error boundary.

If the value is Ok, then throw returns the value, not aborting the rendering process.

The call stack is saved for this component and provided to the error boundary

#[component]
fn App(cx: Scope, count: String) -> Element {
    let id: i32 = count.parse().throw(cx)?;

    cx.render(rsx! {
        div { "Count {}" }
    })
}

Required Associated Types

type Out

The value that will be returned in if the given value is Ok.

Required Methods

fn throw(self, cx: &ScopeState) -> Option<Self::Out>

Returns an option that evaluates to None if there is an error, injecting the error to the nearest error boundary.

If the value is Ok, then throw returns the value, not aborting the rendering process.

The call stack is saved for this component and provided to the error boundary

Note that you can also manually throw errors using the throw method on ScopeState directly, which is what this trait shells out to.

#[component]
fn App(cx: Scope, count: String) -> Element {
    let id: i32 = count.parse().throw(cx)?;

    cx.render(rsx! {
        div { "Count {}" }
    })
}

fn throw_with<D: Debug + 'static>( self, cx: &ScopeState, e: impl FnOnce() -> D ) -> Option<Self::Out>

Returns an option that evaluates to None if there is an error, injecting the error to the nearest error boundary.

If the value is Ok, then throw returns the value, not aborting the rendering process.

The call stack is saved for this component and provided to the error boundary

Note that you can also manually throw errors using the throw method on ScopeState directly, which is what this trait shells out to.

#[component]
fn App(cx: Scope, count: String) -> Element {
    let id: i32 = count.parse().throw(cx)?;

    cx.render(rsx! {
        div { "Count {}" }
    })
}